backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiomysql>=0.2.0",
    "alembic>=1.16.5",
    "cloudinary>=1.44.1",
    "cryptography>=45.0.6",
    "email-validator>=2.3.0",
    "fastapi>=0.116.1",
    "openai>=1.102.0",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.10.1",
    "pymysql>=1.1.2",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "sqlalchemy[asyncio]>=2.0.43",
    "uvicorn>=0.35.0",
]
